% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggPvalue.R
\name{aggPvalue}
\alias{aggPvalue}
\title{Aggregate multiple p-values to a single p-value}
\usage{
aggPvalue(pvals, method = c("Fisher", "Simes"), returnstat = FALSE)
}
\arguments{
\item{pvals}{a vector of p-values.}

\item{method}{either "Fisher" or "Simes" method can be used to aggregate p-values.}

\item{returnstat}{logical. If "TRUE" then return the statistic as well as the p-values.}
}
\description{
Aggregate multiple p-values to a single p-value
}
