% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3D.R
\name{boxplot.normalised}
\alias{boxplot.normalised}
\title{Boxplot of expression distribution in each sample.}
\usage{
\method{boxplot}{normalised}(data.before, data.after, condition,
  sample.name)
}
\arguments{
\item{data.before}{a numeric matrix of expression before normalisation, e.g. raw read counts.}

\item{data.after}{a numeric matrix of expression after normalisation, e.g. \eqn{\log_2}-CPM.}

\item{condition}{a vector of conditions, which is used to distinguish and colour the conditions of
samples.}

\item{sample.name}{a vector of sample names, which is used to name and order the samples in the boxplot.}
}
\value{
a list with two elements: "g1" and "g2", which are \code{ggplot} format boxplots of the data
before and after normalisation.
}
\description{
Boxplot of expression distribution in each sample.
}
