% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition2design.R
\name{condition2design}
\alias{condition2design}
\title{Convert conditions to design matrix for linear regression}
\usage{
condition2design(condition, batch.effect = NULL)
}
\arguments{
\item{batch.effect}{batch effect terms estimated from xxx. Default is \code{NULL}.}

\item{condtion}{a vector of conditions of samples corresponding to the expression dataset.}
}
\value{
A design matrix for linear regression.
}
\description{
Convert conditions to design matrix for linear regression
}
