% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low.expression.filter.R
\name{counts.filter}
\alias{counts.filter}
\title{Filter low expressed transcripts and genes based on read counts}
\usage{
counts.filter(counts, mapping, cpm.cut = 1, sample.n = 3, Log = F)
}
\arguments{
\item{counts}{read counts at transcript level.}

\item{mapping}{a data.frame of transcript-gene name mapping, with first column of transcript list and second column of gene list.}

\item{cpm.cut}{CPM cut-off.}

\item{sample.n}{number of samples cut across all samples.}

\item{Log}{if TRUE, the log2-CPM is used for filters.}
}
\value{
A list of following elements:
    \itemize{
      \item{ trans_high: }{ the expressed transcripts}
      \item{ genes_high: }{ the expressed genes}
      \item{ mapping: }{ the remaining mapping data.frame after filter the low expressed transcripts}
    }
}
\description{
Filter low expressed transcripts and genes based on read counts
}
\details{
Read counts are converted to CPM (count per million reads) to filter low expressed transcripts. 

1) An expressed transcript must have \eqn{\geq} \code{sample.n} with CPM \eqn{\geq} \code{cpm.cut}.

2) An expressed gene must have at least one expressed transcript.
}
