% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2CPM.R
\name{counts2CPM}
\alias{counts2CPM}
\title{Convert read counts to counts per million reads (CPM)}
\usage{
counts2CPM(obj, lib.size = NULL, Log = F)
}
\arguments{
\item{obj}{a read count matrix or a \code{\link{DGEList}} object.}

\item{lib.size}{a numeric vector of library size (sequencing depth) of each sample. If \code{NULL}, library
size is calculated from sum of all reads in a sample.}

\item{Log}{logcial, whether to take \eqn{\log_2} of CPM.}
}
\value{
a numeric matrix of CPM or \eqn{\log_2}-CPM.
}
\description{
Convert read counts to counts per million reads (CPM)
}
\details{
CPM is defined as:
\deqn{\frac{counts}{lib.size}\times 10^6}.
To avoid taking \eqn{\log_2} of zero values, small offsets 0.5 and 1 were added to the numerator and 
denominator, respectively, i.e.
\deqn{\log_2 \frac{counts + 0.5}{lib.size+1}\times 10^6}
}
\examples{
y <- matrix(rnbinom(20,size=1,mu=10),5,4)
counts2CPM(y,Log = T)
counts2CPM(y,Log = F)
}
\seealso{
\code{\link[edgeR]{cpm}}.
}
