% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low.expression.filter.R
\name{CPM.filter}
\alias{CPM.filter}
\title{Filter low expression based on CPM (count per million reads)}
\usage{
CPM.filter(counts, sample.n = 3, cpm.cut = 1, ...)
}
\arguments{
\item{counts}{a data.frame or matrix of read counts.}

\item{sample.n}{number of samples}

\item{cpm.cut}{a numeric value of CPM cut-off}
}
\value{
A vector of logcial values (TRUE/FALSE) to indicate the rows of the input data to keep/filter.
}
\description{
Filter low expression based on CPM (count per million reads)
}
\details{
An expressed target must have \eqn{\geq} \code{sample.n} with CPM \eqn{\geq} \code{cpm.cut}.
}
\examples{
x <- rbind(matrix(rpois(50,lambda=4),5,10),matrix(rpois(50,lambda=0.1),5,10))
rownames(x) <- letters[1:10]
filter.idx <- CPM.filter(x,sample.n = 3,cpm.cut = 1)
##The remaining
x[filter.idx,]
rownames(x)[filter.idx]
}
