% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.error.R
\name{data.error}
\alias{data.error}
\title{Calculate standard deviation or standard errors}
\usage{
data.error(x, error.type = c("stderr", "sd"))
}
\arguments{
\item{x}{a numerical vector}

\item{error.type}{method used to calculate data error. Options include "sd" for standard deviation and "stderr" for standard error.}
}
\value{
data error
}
\description{
Calculate standard deviation or standard errors from a numerical vector.
}
\examples{

set.seed(2000)
x=rnorm(1000,mean=0,sd=1)
##standard error
data.error(x,error.type = 'stderr')
##standard deviation
data.error(x,error.type = 'sd')

}
\seealso{
\code{\link{sd}}
}
