% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.color.hue.R
\name{distinct.color}
\alias{distinct.color}
\title{Generate distinct colours}
\usage{
distinct.color(n)
}
\arguments{
\item{n}{number of colours to generate.}
}
\value{
a vector of \code{n} colours.
}
\description{
Generate distinct colours
}
