% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2mapping.R
\name{fa2mapping}
\alias{fa2mapping}
\title{Generate transcript-gene mapping from transcriptome fasta (.fa) file}
\usage{
fa2mapping(pafile)
}
\arguments{
\item{pafile}{the name of the transcriptome .fa file which the data are to be read from.}
}
\value{
a matrix with first column of transcript names and second column of gene ids.
}
\description{
Generate transcript-gene mapping from transcriptome fasta (.fa) file
}
\details{
Transcript-gene association mapping information will be extracted from the description line of a sequence. 
Transcript names are the sequence names after tag ">" or "transcript" while gene IDs are from the sequence names after tag "gene".
}
