% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.report.R
\name{generate.report}
\alias{generate.report}
\title{Generate html, pdf and word report from Rmd (RMardkown) file}
\usage{
generate.report(input.Rmd, report.folder, type = c("all",
  "html_document", "pdf_document", "word_document"))
}
\arguments{
\item{input.Rmd}{the Rmd file used to generate the report.}

\item{report.folder}{the folder to save the reports.}

\item{type}{the ouput document type. Options are "all", "html_document", "pdf_document" and "word_document".}
}
\value{
the generated html, pdf or word report will be saved into the provided \code{report.folder}.
}
\description{
Generate html, pdf and word report from Rmd (RMardkown) file
}
\examples{
generate.report(input.Rmd = 'DE DAS and DTU analysis report.Rmd',
                report.folder = 'report')
}
