% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run3DApp.R
\name{getExample}
\alias{getExample}
\title{Download example data for 3D RNA-seq analysis}
\usage{
getExample(wd = getwd())
}
\arguments{
\item{wd}{directory to save example data. Default is the project working directory.}
}
\description{
Download example data for 3D RNA-seq analysis
}
\examples{
getExample()
}
