% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.color.hue.R
\name{gg.color.hue}
\alias{gg.color.hue}
\title{Generate HCL colours, which are the same to \code{ggplot} default colour palette}
\usage{
gg.color.hue(n)
}
\arguments{
\item{n}{number of colours to generate.}
}
\value{
a vector of \code{n} colours.
}
\description{
Generate HCL colours, which are the same to \code{ggplot} default colour palette
}
