% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2mapping.R
\name{gtf2mapping}
\alias{gtf2mapping}
\title{Generate transcript-gene mapping from transcriptome gtf file}
\usage{
gtf2mapping(gtfile)
}
\arguments{
\item{gtfile}{the name of the transcriptome gtf file which the data are to be read from.}
}
\value{
a matrix with first column of transcript names and second column of gene ids.
}
\description{
Generate transcript-gene mapping from transcriptome gtf file
}
\details{
Transcript-gene association mapping information will be extracted from the last column of the gtf file. Transcript names
are from the tag "transcript_id" and gene IDs are from the tag "gene_id".
}
