% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low.expression.filter.R
\name{low.expression.filter}
\alias{low.expression.filter}
\title{Filter low expressed transcripts and genes}
\usage{
low.expression.filter(abundance, mapping, abundance.cut = 1,
  sample.n = 3, Log = F, unit = c("counts", "TPM"))
}
\arguments{
\item{abundance}{a data.frame/matrix of read counts/TPMs at transcript level.}

\item{mapping}{a data.frame of transcript-gene name mapping, with first column of transcript list and second column of gene list.}

\item{abundance.cut}{a numeric value of abundance cut-off.}

\item{sample.n}{a number of samples cut across all samples.}

\item{Log}{logical. If TRUE, the log2-CPM is used for filters when read counts are used as abundance.}

\item{unit}{the abundance used to filter the low expressed genes/transcripts. Options are "counts" and "TPM".}
}
\value{
A list of following elements:
    \itemize{
      \item{ trans_high: }{ the expressed transcripts}
      \item{ genes_high: }{ the expressed genes}
      \item{ mapping: }{ the remaining mapping data.frame after filter the low expressed transcripts}
    }
}
\description{
Filter low expressed transcripts and genes
}
\details{
If read counts are used, they are converted to CPM (count per million reads) to filter low expressed transcripts. 

1) An expressed transcript must have \eqn{\geq} \code{sample.n} with CPM \eqn{\geq} \code{cpm.cut}.

2) An expressed gene must have at least one expressed transcript.
}
