% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3D.R
\name{plotEulerDiagram}
\alias{plotEulerDiagram}
\title{Plot Euler diagram}
\usage{
plotEulerDiagram(x, fill = gg.color.hue(length(x)),
  shape = c("ellipse", "circle"), ...)
}
\arguments{
\item{x}{(1) a list of individuals in different sets for comparisons or (2) a vecotr of numbers of set relations.}

\item{fill}{a vector of colours to fill the diagram.}

\item{shape}{geometric shape used in the diagram.}

\item{...}{arguments passed down to the \code{\link{eulerr::euler}} function.}
}
\value{
a Euler diagram
}
\description{
Plot Euler diagram
}
\examples{
x <- letters[1:10]
y <- letters[5:15]
z <- set2(x,y)
combo <- c(x=length(z$x.only),y=length(z$y.only),"x&y"=length(z$xy))
plotEulerDiagram(list(x=x,y=y))
plotEulerDiagram(combo)
}
\seealso{
\code{\link{eulerr::euler}}.
}
