% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDD.flow.chart.R
\name{plotFlowChart}
\alias{plotFlowChart}
\title{Plot flow-chart graph of DE and DAS results}
\usage{
plotFlowChart(expressed, x, y, type = c("genes", "transcripts"),
  pval.cutoff = 0.01, lfc.cutoff = 1, deltaPS.cutoff = 0.1)
}
\arguments{
\item{expressed}{a vector of expressed genes/transcripts.}

\item{x}{a vector of DE genes/transcripts.}

\item{y}{a vector of DAS genes/DTU transcripts.}

\item{type}{a character to indicate the provided list type. Options are "genes" and "transcripts".}

\item{pval.cutoff, lfc.cutoff, deltaPS.cutoff}{the cut-offs used to determine the significant
genes/transcripts.}
}
\value{
a flow-chart plot.
}
\description{
The flow-chart shows the results of DE vs DAS genes or DE vs DTU transcripts.
For example, the expressed genes are divided into DE genes and not DE genes; the DE genes
are divided into DE only (only transcription regulation) and DE&DAS (both transcription
and AS regulation) genes; the not DE genes are divided into DAS only (only AS regulation) and
no expression/AS change (no regulation). The same divisions to DE and/or DTU transcripts.
}
