% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3D.R
\name{plotGO}
\alias{plotGO}
\title{Bar plot of GO annotation}
\usage{
plotGO(go.table, col.idx, plot.title = "GO annotation",
  go.text.cut = 50, legend.position = "right")
}
\arguments{
\item{go.table}{a data.frame of GO annotation table, with first column "Category" (i.e. BP, CC and MF),
second column of Go "Term" and remaining columns with significant statistics, e.g. "Count", "-log10(FDR)", etc.}

\item{col.idx}{a character to indicate which column of statistics to use to report the significance, e.g.
\code{col.idx="-log10(FDR)"}. \code{col.idx} must match to the column names of \code{go.table}.}

\item{plot.title}{the titile to show on the plot.}
}
\value{
a plot in \code{ggplot} format.
}
\description{
Bar plot of GO annotation
}
\examples{
data(go.table)
plotGO(go.table = go.table,col.idx = '-log10(FDR)',plot.title = 'GO annotation: DE genes')

}
