% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.variance.R
\name{plotMeanVariance}
\alias{plotMeanVariance}
\title{Plot the mean-variance trend}
\usage{
plotMeanVariance(
  x,
  y,
  l,
  fit.line.col = "red",
  x.lab = "log2( count size + 0.5 )",
  y.lab = "Sqrt(standard deviation)",
  main = "Mean-variance trend",
  lwd = 1.5,
  ...
)
}
\arguments{
\item{x}{a numeric vector of mean value from the results of \code{\link{trend.mean.variance}}.}

\item{y}{a numeric vector of variance value from the results of \code{\link{trend.mean.variance}}.}

\item{x.lab, y.lab}{a string of x-axis label and y-axis label, respectively.}

\item{main}{plot title.}

\item{...}{additional arguments passed to \code{\link{plot}}.}
}
\value{
a plot
}
\description{
Plot the mean-variance trend
}
