% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.density.R
\name{plotPWISnumber}
\alias{plotPWISnumber}
\title{Barplot of switch numbers in contrast groups}
\usage{
plotPWISnumber(x, angle = 0, hjust = 0.5, vjust = 0.5,
  fill = "dodgerblue3")
}
\arguments{
\item{x}{a data frame with first column of contrast groups and second column of isoform switch numbers in the contrast groups}

\item{fill}{colour to fill the bars}

\item{xlab.angle}{rotate angle of x-axis labels.}

\item{xlab.hjust}{horizontal adjust of x-axis labels.}
}
\description{
Barplot of switch numbers in contrast groups
}
\examples{
x <- data.frame(contrast=c('B-A','C-A'),number=c(500,1000))
plotPWISnumber(x)
}
