% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSIS.R
\name{plotTSIS}
\alias{plotTSIS}
\title{Plot time-series switches of a pair of isoforms}
\usage{
plotTSIS(data2plot, scores = NULL, iso1 = NULL, iso2 = NULL,
  gene.name = NULL, y.lab = "Expression", make.plotly = F, times,
  prob.cutoff = 0.5, x.lower.boundary = 9, x.upper.boundary = 17,
  show.region = T, show.scores = T, error.type = "stderr",
  show.errorbar = T, errorbar.width = 0.03, errorbar.size = 0.5,
  line.width = 1, point.size = 3, marker.size = 1, spline = F,
  spline.df = NULL, ribbon.plot = F)
}
\arguments{
\item{data2plot}{the transcript isoform expression data for switch plots with rows of isoforms and columns of samples. See \code{TSIS.data$data.exp} for details.}

\item{scores}{the output scores of functions \code{\link{iso.switch}} and \code{\link{score.filter}}. Default is NULL. To show score labels on the plot,
the scores must be provided.}

\item{iso1, iso2}{character string names of isoforms to plot.
the input \code{data2plot} must be a data frame of two rows and the row names are used as isoforms to plot.}

\item{gene.name}{a character string of gene name to show as the title of the plot. If \code{gene.name=NULL}, the plot title will be "iso1_vs_iso2".}

\item{y.lab}{the y axis label of the plot, default is "Expression".}

\item{make.plotly}{logical, to plot \code{\link{plotly}} format figures (TRUE) or ggplot2 format figures(FALSE)?. See details in \code{\link{ggplotly}} in \code{\link{plotly}} R pacakge.}

\item{times}{a numeric vector of time labellings of all the relicated samples, e.g. 1,1,1,2,2,2,3,3,3,...}

\item{prob.cutoff}{the cut-off of switch frequencies/probabilities to label the switch points.}

\item{x.lower.boundary, x.upper.boundary}{Specifies the time frame of interest to investigate the isoform switches.}

\item{show.region}{logical, to highlight the time frame under investigation (TRUE) or not (FALSE)?}

\item{show.scores}{logical, to show score labels on the plot (TRUE) or not (FALSE)? The \code{scores} object must be provided.}

\item{error.type}{the error type used to show the error bar in the plots. Options are "stderr" for standard error and "sd" for standard deviation. See details in \code{\link{data.error}}.}

\item{show.errorbar}{logical, to show error bar (TRUE) or not (FALSE) in the error bar plot.}

\item{errorbar.width, errorbar.size}{the width and size of error bars. See detials in \code{\link{geom_errorbar}} in \code{\link{ggplot2}} R pacakge.}

\item{line.width, point.size}{line width and point marker size of the plots.}

\item{spline}{logical, to plot the spline smoothed lines (TRUE) or the lines of mean expression (FALSE).}

\item{spline.df}{the degree of freedom used for spline.  The value must be the same as in the function \code{\link{iso.switch}}. See spline details in \code{\link{geom_smooth}} in \code{\link{ggplot2}}
and \code{\link{ns}} in \code{\link{splines}}.}

\item{ribbon.plot}{logcial, to make ribbon plot (TRUE) or error bar plot (FALSE). See ribbon plot details in \code{\link{geom_smooth}} in \code{\link{ggplot2}} R pacakge.}
}
\value{
a ggplot2 or \code{\link{ggplotly}} (if \code{plotly=TRUE}) plot.
}
\description{
Plot time-series switches of a pair of isoforms
}
\examples{

#generate random datasets
set.seed(100)
data2plot<-data.frame(matrix(abs(rnorm(60)),nrow=2),row.names = c('iso1','iso2'))
#error bar plot
plotTSIS(data2plot=data2plot,scores=NULL,iso1=NULL,iso2=NULL,gene.name=NULL,
        y.lab='Expression',make.plotly=F,times=rep(1:10,each=3),prob.cutoff=0.5,x.lower.boundary=3,
        x.upper.boundary=8,show.region=T,error.type='stderr',ribbon.plot = F)

#ribbon plot
plotTSIS(data2plot=data2plot,scores=NULL,iso1=NULL,iso2=NULL,gene.name=NULL,
         y.lab='Expression',make.plotly=F,times=rep(1:10,each=3),prob.cutoff=0.5,x.lower.boundary=3,
         x.upper.boundary=8,show.region=T,error.type='stderr',ribbon.plot = T)

}
\seealso{
\code{\link{ggplotly}}, \code{\link{iso.switch}}, \code{\link{score.filter}}, \code{\link{data.error}}, \code{\link{geom_smooth}}, \code{\link{ns}}
}
