% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.density.R
\name{plotTSISnumber}
\alias{plotTSISnumber}
\title{Barplot of switch numbers across time-points}
\usage{
plotTSISnumber(
  x,
  time.points,
  angle = 0,
  hjust = 0.5,
  vjust = 0.5,
  fill = "dodgerblue3"
)
}
\arguments{
\item{x}{a numeric vector of isoform switch x-axis coordinates}

\item{time.points}{either a numeric or character vector of time-points, where the switches occur.}

\item{fill}{colour to fill the bars}

\item{xlab.angle}{rotate angle of x-axis labels.}

\item{xlab.hjust}{horizontal adjust of x-axis labels.}
}
\description{
Barplot of switch numbers across time-points
}
\examples{
set.seed(1000)
x <- runif(100,min = 1,max = 3)
plotTSISnumber(x = x,time.points = c('T2','T10','T19'))
plotTSISnumber(x = x,time.points = 1:3)

}
