% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3D.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(
  data2plot,
  xlab,
  ylab,
  title = "Volcano plot",
  col0 = "black",
  col1 = "red",
  size = 1,
  top.n = 10
)
}
\arguments{
\item{data2plot}{a dataframe with a column of "target" names, a column of x coordinates, a column of y coordinates and
a column of "Significant" or "Not significant" to distinguish significance.}

\item{xlab}{the x-axis label.}

\item{ylab}{the y-axis label.}

\item{title}{the title of the plot.}

\item{col0}{colour of not significant points.}

\item{col1}{colour of significant points.}

\item{size}{the point size on the plot.}

\item{top.n}{show target labels of top.n distance to (0,0) coordinate.}
}
\value{
a Volcano diagram
}
\description{
Volcano plot
}
