% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.batch.R
\name{remove.batch}
\alias{remove.batch}
\title{Estimate batch effects between biological replicates}
\usage{
remove.batch(
  read.counts,
  condition,
  design = NULL,
  contrast = NULL,
  group = NULL,
  method = c("RUVr", "RUVg", "RUVs"),
  cIdx = NULL,
  k = 1,
  ...
)
}
\arguments{
\item{read.counts}{data.frame/matrix of read counts.}

\item{condition}{condition labels of the columns in read count data, e.g. c('A','A','A','B','B','B','C','C','C').}

\item{design}{design matrix relating to conditions to be preserved. If \code{NULL}, design matrix will beg
generated from provided \code{condition}.}

\item{contrast}{a vector of contrast groups for condition comparisons, e.g. c('B-A','C-A').}

\item{group}{a vector or factor giving the experimental group/condition for each sample/library. See \code{\link[edgeR]{DGEList}} for details.}

\item{method}{one of "RUVr", "RUVg" and "RUVs". See the RUVSeq R package for details \url{http://bioconductor.org/packages/release/bioc/html/RUVSeq.html}.}

\item{cIdx}{a vector of character, logical or numeric to indicate the subset of genes to 
be used as negative controls in the estimation of the factors of unwanted variation. Default is \code{NULL}.}

\item{k}{a integer number of factors of unwanted variation to be estimated from the data.}

\item{...}{additional arguments passed to \code{\link[edgeR]{DGEList}}.}
}
\value{
A list object with elements:
    \itemize{
      \item{\code{W:} the biological-replicates-by-factors matrix with the estiamted factors of batch effects, 
      which can be passed to the design matrix of linear regression as batch effect terms.}
      \item{\code{normalizedCounts:} a expression matrix the same dimension as the input read counts, in which
      the batch effects have been removed according to the estimated factors. }
      \item{\code{method:} the method used to estimate the batch effects (RUVr, RUVg or RUVs).}
    }
}
\description{
Estimate batch effects between biological replicates
}
\details{
The provided arguments \code{condition}, \code{design}, \code{contrast} and \code{group} are used
to fit a GLM model in \code{\link{edgeR}}, which generates residuals for the \code{\link[RUVSeq]{RUVr}} method.
If the negative controls (not differential expressed genes/transcripts) \code{cIdx=NULL}, the targets with p-value > 0.1
from \code{\link[edgeR]{glmQLFit}} will be used as negative controls. See the packages \code{\link{edgeR}} and \code{\link{RUVSeq}}
for detials.
}
\examples{
samples <- exp.data$samples

##------> sum sequencing replicates
trans_counts <- exp.data$trans_counts
idx <- paste0(samples$condition,'_',samples$brep)
y <- sumarrays(trans_counts,group = idx)

##------> update sample information after sum 
samples <- samples[samples$srep==samples$srep[1],]

##------> normalisation
dge <- DGEList(counts=y)
dge <- calcNormFactors(dge)
data2pca <- t(counts2CPM(obj = dge,Log = T))

##------> plot PCA
groups <- samples$brep
plot.PCA.ind(data2pca = data2pca,dim1 = 'PC1',dim2 = 'PC2',
             groups = groups,ellipse.type='polygon')

##------> remove batch effects
y.new <- remove.batch(read.counts = y,
                      condition = samples$condition,
                      method = 'RUVr')

##------> normalisation and plot PCA again
dge <- DGEList(counts=genes_batch$normalizedCounts)
dge <- suppressWarnings(calcNormFactors(dge))
data2pca <- t(counts2CPM(obj = dge,Log = T))
plot.PCA.ind(data2pca = data2pca,dim1 = 'PC1',dim2 = 'PC2',
             groups = groups,ellipse.type='polygon')
}
