% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.cluster.R
\name{reorderClusters}
\alias{reorderClusters}
\title{Reorder the clusters according to their average exprssion}
\usage{
reorderClusters(clusters, dat)
}
\arguments{
\item{clusters}{a vector of clusters of targets}

\item{data}{expression data of these target}
}
\value{
a vector of ordered clusters
}
\description{
Reorder the clusters according to their average exprssion
}
