% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDf.R
\name{roundDF}
\alias{roundDF}
\title{Round the numeric columns of a data frame}
\usage{
roundDF(df, digits)
}
\arguments{
\item{df}{a data fram object}

\item{digits}{an integer indicating the number of decimal places}

\item{...}{arguments to be passed to \code{\link{round}}.}
}
\description{
Round the numeric columns of a data frame
}
\seealso{
\code{\link{round}}.
}
