% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowratio.R
\name{rowratio}
\alias{rowratio}
\title{Calculate column raito of a matrix or data frame based on a grouping variable}
\usage{
rowratio(x, group, reorder = F, na.rm = T)
}
\arguments{
\item{x}{a matrix or data frame.}

\item{group}{a vector of factor giving grouping, with one element per row of x.}

\item{reorder}{if TRUE, then the result will be in order of row names of x. If row names of x is null, the
results is not reordered.}

\item{na.rm}{logical (TRUE or FALSE). Should NA (including NaN) values be replaced by value 0?}
}
\value{
\code{rowratio} returns a matrix or data frame containing the ratios. There
will be one row per unique value of group.
}
\description{
Compute column ratio across rows of a numeric matrix-like object for each level
of a grouping variable.
}
\examples{
x <- matrix(runif(50), ncol = 5)
group <- sample(1:4, 10, TRUE)
xratio <- rowratio(x, group)

}
\seealso{
\code{\link{rowsum}}, \code{\link{rowmean}}
}
