% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run3DApp.R
\name{run3DApp}
\alias{run3DApp}
\title{Run 3D RNA-seq App}
\usage{
run3DApp(wd = getwd(), ...)
}
\arguments{
\item{wd}{directory to run the App. Default is the project working directory.}

\item{...}{additional arguments passed to \code{runApp} in \code{\link{shiny}} R package.}
}
\description{
Run 3D RNA-seq App
}
\examples{
run3DApp()
}
