% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.filter.R
\name{score.filter}
\alias{score.filter}
\title{Filtering the scores for isoform switch}
\usage{
score.filter(scores, prob.cutoff = 0.5, diff.cutoff = 1,
  t.points.cutoff = 2, pval.cutoff = 0.01, FDR.cutoff = 0.01,
  cor.cutoff = 0.5, data.exp = NULL, mapping = NULL,
  sub.isoform.list = NULL, sub.isoform = F, max.ratio = F,
  x.value.limit = c(9, 17))
}
\arguments{
\item{scores}{the scores object output from \code{\link{iso.switch}}.}

\item{prob.cutoff, diff.cutoff, t.points.cutoff, pval.cutoff, FDR.cutoff, cor.cutoff}{the cut-offs corresponding to switch frequencies/probablities,
sum of average sample differences, p-value, adjusted p-value (FDR) and time points cut-offs for both intervals before and after switch and Pearson correlation.}

\item{data.exp, mapping}{the expression and gene-isoform mapping data.}

\item{sub.isoform.list}{a vector of isoforms to output the corresponding results (see \code{TSIS.data$sub.isoforms}).}

\item{sub.isoform}{logical, to output subset of the results(TRUE) or not (FALSE). If TRUE, \code{sub.isoform.list} must be provided.}

\item{max.ratio}{logical, to show maximum abundant isoform results(TRUE) or not (FALSE). If TRUE, data.exp and mapping data must be
provided to calculate the isoform ratios to the genes using  \code{\link{rowratio}}.}

\item{x.value.limit}{the region of x axis (time) for investigation.}
}
\value{
a table of scores after filtering.
}
\description{
Filtering the scores output from \code{\link{iso.switch}}.
}
\details{
Users can set cut-offs, such as for the probability/frequency of switch and sum of average differences, to further refine the switch results.
}
