% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarrays.R
\name{sumarrays}
\alias{sumarrays}
\title{Sum Over Replicate Arrays}
\usage{
sumarrays(x, group = NULL)
}
\arguments{
\item{x}{a matrix-like object.}

\item{group}{grouping of sample identifier.}
}
\value{
A data object of the same class as x with a column for sums according grouping.
}
\description{
Sum Over Replicate Arrays
}
\examples{
set.seed(100)
x<- matrix(rnorm(8*4),8,4)
colnames(x) <- c("a","a","b","b")
sumarrays(x)
data.frame(a=x[,1]+x[,2],b=x[,3]+x[,4])
}
