% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DDD.R
\name{summary.DE.target}
\alias{summary.DE.target}
\title{Summary DE genes and transcripts}
\usage{
\method{summary}{DE.target}(stat, cutoff = c(adj.pval = 0.01, log2FC =
  1))
}
\arguments{
\item{stat}{a data.frame object with first column of "target", second column of "contrast",
third and fourth columns of DE gene/transcript statistics (i.e. "adj.pval" and "log2FC", respectively).}

\item{cutoff}{a numeric vector of cut-offs to "adj.pval" and "log2FC", e.g. \code{cutoff=c(adj.pval=0.01,log2FC=1)}.}
}
\value{
a data.frame object, which is a subset of input \code{stat} after applying the \code{cutoff}.
}
\description{
Summary DE genes and transcripts
}
