% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.3D.R
\name{summaryDAStarget}
\alias{summaryDAStarget}
\title{Summary DAS genes and DTU transcripts}
\usage{
summaryDAStarget(stat, lfc, cutoff = c(adj.pval = 0.01, deltaPS = 0.1))
}
\arguments{
\item{stat}{a data.frame object with first column of "target", second column of "contrast",
third and fourth columns of DAS gene/DTU transcript statistics (i.e. "adj.pval" and "maxdeltaPS"/"deltaPS", respectively).}

\item{cutoff}{a numeric vector of cut-offs for the statistics.}
}
\value{
a data.frame object, which is a subset of input \code{stat} after applying the \code{cutoff}.
}
\description{
Summary DAS genes and DTU transcripts
}
