% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.3D.R
\name{summaryDEtarget}
\alias{summaryDEtarget}
\title{Summary DE genes and transcripts}
\usage{
summaryDEtarget(stat, cutoff = c(adj.pval = 0.01, log2FC = 1))
}
\arguments{
\item{stat}{a data.frame object with first column of "target", second column of "contrast",
third and fourth columns of DE gene/transcript statistics (i.e. "adj.pval" and "log2FC", respectively).}

\item{cutoff}{a numeric vector of cut-offs to "adj.pval" and "log2FC".}
}
\value{
a data.frame object, which is a subset of input \code{stat} after applying the \code{cutoff}.
}
\description{
Summary DE genes and transcripts
}
