% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.3D.R
\name{summaryStat}
\alias{summaryStat}
\title{Merge two testing statistics}
\usage{
summaryStat(
  x,
  y,
  target,
  contrast = NULL,
  stat.type = c("adj.pval", "lfc"),
  srot.by = stat.type[1]
)
}
\arguments{
\item{x, y}{data.frame object of statistics (e.g. p-values, log2-fold changes, deltaPS), with rows of targets and
columns statistics in contrast groups.}

\item{contrast}{a vector of contrast groups, e.g. \code{contrast = c('B-A','C-A')}. If \code{NULL}, the column names of
\code{x} are used as \code{contrast}.}

\item{stat.type}{a vector of statistic types of \code{x} and \code{y}, respectively, e.g.
\code{stat.type = c('adj.pval','lfc')}. \code{stat.type} is passed to third and fourth column names of the output data.frame.}

\item{srot.by}{a column name to sort the output data.frame.}
}
\value{
\code{summaryStat} returns a data.frame object with first column of target, second column of contrast groups, third column name of
statistics \code{x} and fourth column of statistic \code{y}.
}
\description{
Merge two testing statistics
}
