% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transAbundance2PS.R
\name{transAbundance2PS}
\alias{transAbundance2PS}
\title{Calculate transcript PS (percent spliced) and \eqn{\Delta}PS from abundance}
\usage{
transAbundance2PS(transAbundance = NULL, PS = NULL, contrast,
  condition, mapping)
}
\arguments{
\item{transAbundance}{matrix of transcript level abundance, e.g. TPM and read count exprssion.}

\item{PS}{PS value matrix. If a PS matrix provided, the \eqn{\Delta}PS values is directly calculated from the PS matrix 
based on the contrast groups, otherwise, new PS value will be generate. Default is \code{NULL}.}

\item{contrast}{a vector of contrast groups to calculate \eqn{\Delta}PS values, e.g. c('B-A','C-A'), which compares conditions 
"B" and "C" to condition "A".}

\item{condition}{a vector of condition labels which match the columns in the abundance matrix, 
e.g. c('A','A','A','B','B','B','C','C','C'), which means the abundance matrix corresponds to 3 conditions and each condition 
has 3 replicates.}

\item{mapping}{transcript-gene mapping matrix. First column is transcript list with column name "TXNAME" and 
second column is gene list with column name "GENEID".}
}
\value{
a list of two data.frames: "PS" and "deltaPS".
}
\description{
Calculate transcript PS (percent spliced) and \eqn{\Delta}PS from abundance
}
\details{
PS is defined as the ratio of transcript abundance divided by the gene abundance 
(sum of all transcript abundance of the same gene).
}
\examples{
set.seed(36)
transAbundance <- matrix(rnorm(36,mean=10,sd=1),ncol = 9,nrow = 3)
PS <- NULL
contrast <- c('B-A','C-A')
condition <- c('A','A','A','B','B','B','C','C','C')
mapping <- data.frame(TXNAME=c('trans1.1','trans1.2','trans2'),
                      GENEID=c('gene1','gene1','gene2'))
transAbundance2PS(transAbundance = transAbundance,PS = PS,
                  contrast = contrast,condition = condition,mapping = mapping)
                 
}
