% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.variance.R
\name{trend.mean.variance}
\alias{trend.mean.variance}
\title{Estimate mean-variance trend of expression}
\usage{
trend.mean.variance(obj, design, lib.size = NULL, span = 0.5, ...)
}
\arguments{
\item{obj}{a matrix of gene/transcript read counts, or a \code{\link[edgeR]{DGEList}} object. Read counts
must be non-negative and NAs are not permitted.}

\item{design}{design matrix with rows of samples and columns of conditions to fit a linear model.}

\item{lib.size}{a vector of library size for each sample. If \code{NULL}, the library size will be either extracted
from the \code{DGEList} object or calculated by summing up all the reads of genes/transcripts in each sample.}

\item{span}{a numeric value passed to \code{\link{lowess}} smoothing window as a proportion to fit the mean-variance trend.}

\item{...}{additional arguments passed to \code{\link[limma]{lmFit}} function.}
}
\value{
a list object with elements:
    \itemize{
      \item{ fit: }{ the linear regression results of expression based on design matrix.}
      \item{ sx: }{ a numeric vector of mean values. }
      \item{ sy: }{ a numeric vector of variance values.}
      \item{ l: }{ the \code{lowess} fit result based on sx and sy.}
    }
}
\description{
Estimate mean-variance trend of expression
}
\seealso{
\code{\link[limma]{voom}}, \code{\link{lowess}} and \code{\link{condition2design}}.
}
