% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.intersection.R
\name{ts.intersection}
\alias{ts.intersection}
\title{Searching for intersection points of two time-series}
\usage{
ts.intersection(x1, x2, times)
}
\arguments{
\item{x1}{a vector of values for first time-series}

\item{x2}{a vector of values for second time-series}
}
\value{
a data frame of intersection points. The first and second columns are x and y coordinates values of intersections, respectively.
}
\description{
Searching for intersection points of two time-series
}
