% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.spline.R
\name{ts.spline}
\alias{ts.spline}
\title{Fit a time-series with spline curve}
\usage{
ts.spline(x, times, df = 5, ...)
}
\arguments{
\item{x}{a vector of time-series expression}

\item{times}{a numeric vector of time labellings of all the relicated samples, e.g. 1,1,1,2,2,2,3,3,3,...}

\item{df}{degree of freedom used in \code{\link{ns}} in \code{\link{splines}} package.}

\item{...}{additional arguments passed to \code{\link{predict}}.}

\item{nrep}{number of replicates.}
}
\value{
predictions results returned from \code{\link{predict}}.
}
\description{
Fit a time-series with spline curve
}
\seealso{
\code{\link{lm}}, \code{\link{ns}}, \code{\link{predict}}
}
