% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{TSgroup2contrast}
\alias{TSgroup2contrast}
\title{Generate contrast groups of time-series trend test}
\usage{
TSgroup2contrast(Group, Time, Intercept = list(Time = NULL, Group =
  NULL))
}
\description{
Generate contrast groups of time-series trend test
}
