% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.3D.R
\name{DEvsDAS}
\alias{DEvsDAS}
\alias{DEvsDTU}
\alias{summary3Dnumber}
\title{Compare DE and DAS results}
\usage{
DEvsDAS(DE_genes, DAS_genes, contrast, consensus = F)

DEvsDTU(DE_trans, DTU_trans, contrast, consensus = F)

summary3Dnumber(DE_genes, DAS_genes, DE_trans, DTU_trans, contrast,
  consensus = F)
}
\arguments{
\item{DE_genes, DAS_genes, DE_trans, DTU_trans}{data.frame objects of DE genes, DAS genes, DE transcripts and DTU transcripts,
which are the outputs of \code{\link{summaryDEtarget}} and \code{\link{summaryDAStarget}}.}

\item{contrast}{a vector of contrast groups, e.g. \code{contrast = c('B-A','C-A')}, which compares condition B and C to 
condition A.}
}
\value{
a data.frame with first column of contrast goups, second column of DE only genes/transcripts, third column of
DE&DAS genes or DE&DTU transcripts and fourth column of DAS only genes or DTU only transcripts.
}
\description{
Compare DE and DAS results
}
\details{
In each contrast group, \code{DEvsDAS} compares the DE and DAS genes; \code{DEvsDTU} compares the DE and DTU
transcripts; and \code{summary3Dnumber} summarises the DE/DAS/DTU gene/transcript numbers.
}
