% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2dataframe.R
\name{list2dataframe}
\alias{list2dataframe}
\title{Convert list to data frame}
\usage{
list2dataframe(x)
}
\arguments{
\item{x}{a list object with elements of vectors. The lengths of these vectors can be different.}
}
\value{
a data frame.
}
\description{
Generate data frame from a list object of vectors. If the lengths of the vectors in the list object are different, NA is filled in the data frame.
}
