% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3D.R
\name{plotAbundance}
\alias{plotAbundance}
\alias{plotPS}
\title{Plot gene and transcript expression profiles}
\usage{
plotAbundance(data.exp, gene, mapping, genes.ann = NULL,
  trans.ann = NULL, trans.expressed = NULL, reps, groups = NULL,
  sliceProfile = F, x.lab = "Conditions", y.lab = "TPM",
  marker.size = 3, legend.text.size = 11, error.type = "stderr",
  error.bar = T, show.annotation = T, plot.gene = T)

plotPS(data.exp, gene, mapping, genes.ann = NULL, trans.ann = NULL,
  trans.expressed = NULL, reps, groups = NULL, sliceProfile = F,
  y.lab = "TPM", x.lab = "Conditions", marker.size = 3,
  legend.text.size = 11, show.annotation = T)
}
\arguments{
\item{data.exp}{a data.frame of transcript level expression, e.g. read counts and TPM.}

\item{gene}{a gene name to plot.}

\item{mapping}{a data.frame of transcript-gene mapping (first column is transcript list and second column is gene list).}

\item{genes.ann}{a data.frame of gene annotation with first column "target" and second column "annotation". Default is \code{NULL}.}

\item{trans.ann}{a data.frame of transcript annotation with first column "target" and second column "annotation".
Default is \code{NULL}.}

\item{trans.expressed}{a vector of expressed transcripts. If not \code{NULL}, only the profiles of provided transcripts
will be shown in the plot.}

\item{reps}{a vecotr of replicate labels, which provide the grouping information of biolocial replciates to 
calculate the average expression in each conditions.}

\item{groups}{a vecotr of grouping labels for \code{reps}, which provide the grouping information to slice the profile plot into different segments, e.g. samples in "treatment"
vs "Control", samples in different development stages, etc. The vector length of \code{groups} is the same to the length of \code{reps}.}

\item{sliceProfile}{logical, whether to slice profile plot according to \code{groups} labels.}

\item{x.lab, y.lab}{characters for x-axis and y-axis labels, respectively.}

\item{marker.size}{a number passed to \code{geom_point} to control the point size in the plot.}

\item{legend.text.size}{a number to control the legend text size.}

\item{error.type}{error type to make the error bars. Options are: "stderr" for standard error and
"sd" for standard deviation.}

\item{error.bar}{logical, whether to show error bars on the profile plot.}

\item{show.annotation}{logical, whether to show the annotations provided in \code{genes.ann} and \code{trans.ann} on the plot.}

\item{plot.gene}{logical, whether to show the gene level expression on the plot. The gene expression is the total of all the
transcript expression.}
}
\value{
a plot in \code{ggplot} format.
}
\description{
Plot gene and transcript expression profiles
}
\details{
\code{plotAbundance} is used to plot the gene and/or trnascript level abundance while \code{plotPS} is used to plot
the percent spliced (PS) of transcript. PS is defined as the ratio of transcript expression to the total of all transcript
expression (i.e. gene expression) in the same gene.
}
\examples{
data(exp.data)
plotAbundance(data.exp = exp.data$trans_TPM,
              gene = 'AT1G01020',
              mapping = exp.data$mapping,
              genes.ann = exp.data$genes.ann,
              trans.ann = exp.data$trans.ann,
              reps = exp.data$samples$condition)

plotPS(data.exp = exp.data$trans_TPM,
              gene = 'AT1G01020',
              mapping = exp.data$mapping,
              genes.ann = exp.data$genes.ann,
              trans.ann = exp.data$trans.ann,
              reps = exp.data$samples$condition)
}
\seealso{
\code{\link{data.error}}
}
