% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeDRNAseq.app.R
\name{ThreeDRNAseq.app}
\alias{ThreeDRNAseq.app}
\title{The shiny app for 3D analysis}
\usage{
ThreeDRNAseq.app(data.size.max = 300)
}
\arguments{
\item{data.size.max}{maximum size limit for unload files in Shiny. Default is 300 (MB).}
}
\value{
the Shiny App.
}
\description{
The shiny app for 3D analysis
}
\examples{
RNAseq3D.app()

}
\seealso{
\code{\link{shiny}}
}
